refactor(#1404): extract CLI ACK-provider construction into an intentional internal module
otReviewAgent follow-up: don't widen the publisher-runner surface just to
test factory wiring. Move `createV10ACKProviderForPublisher` + the
`ACKTransportFactory` interface out of `publisher-runner.ts` into a new
dedicated module `packages/cli/src/ack-provider.ts` with a stable,
deliberately-importable internal contract. `publisher-runner.ts` now only
IMPORTS them (no longer defines/exports a private helper for tests), and
the wiring test imports from `ack-provider.ts` — a deliberate internal
seam, not re-exported from the package entry point. Pure refactor:
full build green; full CLI suite 2260 pass / 0 fail.
